-added moving blocks

+++ src/com/hshedges/game/objects/Block.java
@@ -6,7 +6,7 @@ import com.hshedges.game.states.GameState;
 import java.awt.*;
 
 public class Block extends Rectangle {
-    public int x,y;
+    public double x,y;
     public int id;
     public static final int WIDTH = 128;
     public static final int HEIGHT = 128;
@@ -22,7 +22,7 @@ public class Block extends Rectangle {
         int xOff = (int)GameState.xOffset;
         int yOff = (int)GameState.yOffset;
         if(id != 0){
-            g.drawImage(Images.blocks[id-1],x - xOff ,y - yOff ,WIDTH,HEIGHT,null);
+            g.drawImage(Images.blocks[id-1],(int)x - xOff ,(int)y - yOff ,WIDTH,HEIGHT,null);
         }
     }
 }
